refactor

+++ src/main/java/com/gabrielcsilva1/Portal_Egresso/domain/services/PositionService.java
@@ -1,18 +1,20 @@
 package com.gabrielcsilva1.Portal_Egresso.domain.services;
 
+import java.util.List;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.gabrielcsilva1.Portal_Egresso.domain.dtos.PositionDTO;
-import com.gabrielcsilva1.Portal_Egresso.domain.dtos.position.UpdatePositionDTO;
 import com.gabrielcsilva1.Portal_Egresso.domain.entities.Position;
 import com.gabrielcsilva1.Portal_Egresso.domain.repositories.GraduateRepository;
 import com.gabrielcsilva1.Portal_Egresso.domain.repositories.PositionRepository;
-import com.gabrielcsilva1.Portal_Egresso.domain.services.exeptions.GraduateNotFoundException;
-import com.gabrielcsilva1.Portal_Egresso.domain.services.exeptions.InvalidEndYearException;
-import com.gabrielcsilva1.Portal_Egresso.domain.services.exeptions.ResourceNotFoundException;
+import com.gabrielcsilva1.Portal_Egresso.dtos.request.position.RequestCreatePositionJson;
+import com.gabrielcsilva1.Portal_Egresso.dtos.request.position.RequestUpdatePositionJson;
+import com.gabrielcsilva1.Portal_Egresso.exeptions.GraduateNotFoundException;
+import com.gabrielcsilva1.Portal_Egresso.exeptions.InvalidEndYearException;
+import com.gabrielcsilva1.Portal_Egresso.exeptions.ResourceNotFoundException;
+import com.gabrielcsilva1.Portal_Egresso.exeptions.core.BadRequestException;
 
 @Service
 public class PositionService {
@@ -22,8 +24,8 @@ public class PositionService {
   @Autowired
   private GraduateRepository graduateRepository;
   
-  public Position registerGraduatePosition(PositionDTO positionDTO) {
-    var graduate = this.graduateRepository.findById(positionDTO.getGraduateId());
+  public Position registerGraduatePosition(UUID graduateId, RequestCreatePositionJson positionDTO) {
+    var graduate = this.graduateRepository.findById(graduateId);
 
     if (graduate.isEmpty()) {
       throw new GraduateNotFoundException();
@@ -50,9 +52,14 @@ public class PositionService {
     return this.positionRepository.save(position);
   }
 
-  public Position updateGraduatePosition(UUID positionId, UpdatePositionDTO positionDTO) {
+  public Position updateGraduatePosition(UUID positionId, UUID graduateId, RequestUpdatePositionJson positionDTO) {
     Position position = this.positionRepository.findById(positionId)
       .orElseThrow(() -> new ResourceNotFoundException());
+
+    boolean isSameGraduateId = position.getGraduate().getId().equals(graduateId);
+    if (isSameGraduateId == false) {
+      throw new BadRequestException();
+    }
 
     position.setDescription(positionDTO.getDescription());
     position.setLocation(positionDTO.getLocation());
@@ -62,10 +69,19 @@ public class PositionService {
     return this.positionRepository.save(position);
   }
 
-  public void deleteGraduatePosition(UUID positionId) {
+  public void deleteGraduatePosition(UUID positionId, UUID graduateId) {
     Position position = this.positionRepository.findById(positionId)
       .orElseThrow(() -> new ResourceNotFoundException());
 
+    boolean isSameGraduateId = position.getGraduate().getId().equals(graduateId);
+    if (isSameGraduateId == false) {
+      throw new BadRequestException();
+    }
+
     this.positionRepository.delete(position);
+  }
+
+  public List<Position> fetchGraduatePositions(UUID graduateId) {
+    return this.positionRepository.findByGraduateId(graduateId);
   }
 }
